feat: [UEPR-187] [UEPR-192] [UEPR-194] renaming, reordering components and changes in logic
