+++ src/components/modal/terms-of-use/terms-of-use-modal.scss
@@ -1,5 +1,5 @@
-@import "../../colors";
-@import "../../frameless";
+@import "../../../colors";
+@import "../../../frameless";
 
 .tou-modal-overlay {
     background-color: transparentize($overlay-gray, 0.2);
