--- src/components/tou-modal/tou-banner.jsx
@@ -1,44 +1,0 @@
-import React, {useState} from 'react';
-import {injectIntl, FormattedMessage} from 'react-intl';
-import TosModalReminderUnder16 from './variations/tou-reminder-under-16.jsx';
-
-require('./tou-modal.scss');
-
-const TermsOfUseLink = chunks => {
-    const [isModalVisible, setIsModalVisible] = useState(false);
-    return (<>
-        <a
-            className="tou-banner-link"
-            // eslint-disable-next-line react/jsx-no-bind
-            onClick={() => {
-                setIsModalVisible(true);
-            }}
-        >
-            {chunks}
-        </a>
-        {isModalVisible && <TosModalReminderUnder16
-            // eslint-disable-next-line react/jsx-no-bind
-            onClose={() => {
-                setIsModalVisible(false);
-            }}
-            isOpen={isModalVisible}
-        /> }
-    </>);
-
-};
-
-// eslint-disable-next-line arrow-body-style
-const TermsOfUseBanner = () => {
-    return (
-        <div className="tou-banner">
-            <FormattedMessage
-                id="tos.banner.parentAgreement"
-                values={{
-                    a: TermsOfUseLink
-                }}
-            />
-        </div>
-    );
-};
-
-export default injectIntl(TermsOfUseBanner);
